[Del]remove usage of pylint
