--- .flake8
@@ -1,11 +1,0 @@
-[flake8]
-exclude = .git,__pycache__,clang,build,dist,generated_files,templates,__init__.py,
-ignore =
-    E501 line too long, fixed by black
-    W503 line break before binary operator
-
-    W293 blank line contains whitespace
-    W291 trailing whitespace
-
-[pycodestyle]
-max_line_length = 79
