+++ .travis.yml
@@ -1,11 +1,8 @@
 language: python
 python:
-  - "3.6"
+  - "3.7"
 install:
   - pip install -r requirements.txt
-  - pip install -r requirements-dev.txt
-  - pip install -r tools/ci/requirements.txt
-# command to run tests
+  - pip install flake8
 script:
-  - python tools/ci/check_all.py
-  - python tests/test_all.py
+  - python check.py
