--- requirements-dev.txt
@@ -1,3 +1,0 @@
-pylint
-autopep8
-flake8
